Test Helpers: Move getUserDataDir from nsi_helper.js to spectron_helper.js; Initialize db only once in db_helper.initDatabase
